requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
requests
